test(opencode): keep both halves of the prompt contract in one suite
The opencode folder carries one test file per subsystem and was already at
five; a sixth for the CLI instruction file would have fragmented the very
contract it tests. The chat runtime sends ADE's instructions on the prompt's
first-class `system` field and a CLI launch reads the same text from an
instruction file, so both now sit in openCodeRuntime.test.ts where the
parity assertion between them can't drift from either side.

No assertions were dropped in the move.
