+++ apps/desktop/src/main/services/opencode/openCodeRuntime.test.ts
@@ -1,4 +1,7 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
+import fs from "node:fs";
+import os from "node:os";
+import path from "node:path";
 
 const mockState = vi.hoisted(() => {
   let nextSessionId = 1;
@@ -125,6 +128,12 @@ import {
   acquireDedicatedOpenCodeServer,
   acquireSharedOpenCodeServer,
 } from "./openCodeServerManager";
+import { buildCodingAgentSystemPrompt } from "../ai/tools/systemPrompt";
+import {
+  buildOpenCodeAdeInstructions,
+  ensureOpenCodeAdeInstructionsFile,
+  openCodeAdeInstructionsPath,
+} from "./openCodeAdeInstructions";
 
 /** The flat parameters object from the most recent v2 `session.promptAsync`. */
 function openCodePromptParams(): Record<string, unknown> {
@@ -599,5 +608,91 @@ describe("buildOpenCodeConfig user-owned keys", () => {
     // opencode.json, repointing a configured remote host back at localhost.
     expect(config().provider ?? {}).not.toHaveProperty("ollama");
     expect(config().provider ?? {}).not.toHaveProperty("lmstudio");
+  });
+});
+
+// The other half of the OpenCode prompt contract. The chat runtime above sends
+// ADE's instructions on the prompt's first-class `system` field; a CLI launch has
+// no per-request system channel, so it reads the same text from an instruction
+// file referenced by config. Both live here so the two transports cannot drift.
+describe("ADE instructions for the tracked OpenCode CLI", () => {
+  let projectRoot: string;
+  let lane: string;
+
+  beforeEach(() => {
+    projectRoot = fs.mkdtempSync(path.join(os.tmpdir(), "ade-oc-instr-"));
+    lane = path.join(projectRoot, ".ade", "worktrees", "lane-1");
+  });
+
+  afterEach(() => {
+    fs.rmSync(projectRoot, { recursive: true, force: true });
+  });
+
+  it("gives chat and the tracked CLI the same ADE base prompt", () => {
+    // The contract is "one slim ADE base prompt", not "byte-identical
+    // transport": chat sends it through the SDK's `system` field while the CLI
+    // gets it as an instruction file. What must not drift is the instruction
+    // text itself, so the CLI file has to contain exactly what the chat runtime
+    // builds for the same lane and permission mode.
+    const chatPrompt = buildCodingAgentSystemPrompt({
+      cwd: lane,
+      mode: "coding",
+      permissionMode: "edit",
+      interactive: true,
+      runtime: "opencode",
+    });
+
+    expect(buildOpenCodeAdeInstructions({ laneWorktreePath: lane, permissionMode: "edit" }))
+      .toContain(chatPrompt);
+  });
+
+  it("tracks the permission mode the CLI actually launched with", () => {
+    const plan = buildOpenCodeAdeInstructions({ laneWorktreePath: lane, permissionMode: "plan" });
+    const fullAuto = buildOpenCodeAdeInstructions({ laneWorktreePath: lane, permissionMode: "full-auto" });
+
+    expect(plan).toContain("Plan mode. Stay read-only");
+    expect(fullAuto).toContain("Autonomous mode.");
+    expect(plan).not.toContain("Autonomous mode.");
+  });
+
+  it("writes into ADE's own cache, never the lane worktree or a shared temp dir", () => {
+    const written = ensureOpenCodeAdeInstructionsFile({ projectRoot, laneWorktreePath: lane, permissionMode: "edit" });
+
+    expect(written).not.toBeNull();
+    expect(written?.startsWith(lane)).toBe(false);
+    expect(written?.startsWith(path.join(projectRoot, ".ade", "cache"))).toBe(true);
+    expect(fs.readFileSync(written!, "utf8")).toContain("ADE's software engineering agent");
+    // A predictable path in a world-writable directory is a symlink target.
+    expect(written?.startsWith(os.tmpdir() + path.sep + "ade-opencode")).toBe(false);
+  });
+
+  it("reuses one stable path per lane and rewrites it on relaunch", () => {
+    const first = ensureOpenCodeAdeInstructionsFile({ projectRoot, laneWorktreePath: lane, permissionMode: "plan" });
+    const second = ensureOpenCodeAdeInstructionsFile({ projectRoot, laneWorktreePath: lane, permissionMode: "full-auto" });
+
+    // A resume must not resurrect the previous launch's permission mode.
+    expect(second).toBe(first);
+    expect(fs.readFileSync(second!, "utf8")).toContain("Autonomous mode.");
+
+    expect(openCodeAdeInstructionsPath({ projectRoot, laneWorktreePath: `${lane}-2` })).not.toBe(first);
+  });
+
+  it("produces a filename OpenCode's basename glob cannot misread", () => {
+    // OpenCode resolves an absolute `instructions` entry by globbing the
+    // basename inside its parent directory, so a glob metacharacter in the name
+    // would match the wrong file or none.
+    const name = path.basename(openCodeAdeInstructionsPath({
+      projectRoot,
+      laneWorktreePath: "/repo/lane with spaces/*/weird?",
+    }));
+    expect(name).toMatch(/^ade-[0-9a-f]{16}\.md$/);
+  });
+
+  it("returns null instead of failing the launch when there is no lane", () => {
+    expect(ensureOpenCodeAdeInstructionsFile({
+      projectRoot,
+      laneWorktreePath: "  ",
+      permissionMode: "edit",
+    })).toBeNull();
   });
 });
